thredion-engine/requirements.txt
fastapi
uvicorn
supabase
faster-whisper
openai
yt-dlp
httpx
beautifulsoup4
readability-lxml
pydantic
python-dotenv
python-multipart
python-jose
Passlib[bcrypt]
sqlalchemy
psycopg2-binary
sentence-transformers
